chore: use fastpool as object pool impl  (#6820)
* chore: use fastpool as object pool impl


* for redis


* for sftp


* for memcached


* run format


* fixup compile


* Revert "run format"

This reverts commit 052cb4cca1cf49a0577dcfbba3c641fe6637a704.

---------

+++ core/src/services/memcached/config.rs
@@ -46,7 +46,7 @@ pub struct MemcachedConfig {
     /// The maximum number of connections allowed.
     ///
     /// default is 10
-    pub connection_pool_max_size: Option<u32>,
+    pub connection_pool_max_size: Option<usize>,
 }
 
 impl Debug for MemcachedConfig {
